requirements.txt
Django>=5.0
psycopg2-binary
dj-database-url
whitenoise[brotli]
cloudinary
django-cloudinary-storage
gunicorn
python-dotenv
django-polymorphic
Pillow
